Move extract_icon to a more appropriate directory
./utils contains stand alone utility scripts; utility routines used
by wine seem to be in lutris/util/wine, so lets move this there.

I am hoping this will address issue #4529
