+++ lutris/util/wine/extract_icon.py
